requirements.txt
python-dotenv
requests
numpy
pandas
scipy
torch
jupyterlab
ipywidgets
matplotlib
scikit-learn
chromadb
jupyter-dash
sentencepiece
pyarrow
faiss-cpu
beautifulsoup4
plotly
bitsandbytes
transformers
sentence-transformers
datasets
accelerate
openai
anthropic
google-generativeai
gradio
gensim
modal
ollama
psutil
sacremoses
setuptools
speedtest-cli
langchain
langchain-core
langchain-text-splitters
langchain-openai
langchain-chroma
langchain-community
feedparser
twilio
pydub
flask
flask-session
gunicorn
rdkit
deepeval
